% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_models.R
\name{consolidate_models}
\alias{consolidate_models}
\title{Consolidate model from model grid}
\usage{
consolidate_models(shared_settings, models)
}
\arguments{
\item{shared_settings}{\code{list} settings that are shared across all models.}

\item{models}{\code{list} with the individual specifications for models in
a model grid.}
}
\value{
\code{list} with a complete model specification, that can be trained
using the 'caret' package.
}
\description{
Consolidate models from shared settings and model specific settings. In case
of any overlap between the two, the model specific settings will apply.
}
