% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_grid.R
\name{model_grid}
\alias{model_grid}
\alias{set_shared_settings}
\alias{add_model}
\alias{consolidate_models}
\alias{xtrain}
\alias{edit_model}
\alias{remove_model}
\title{Create a caret model grid}
\usage{
model_grid()

set_shared_settings(model_grid, ...)

add_model(model_grid, model_name = NULL, custom_control = NULL, ...)

consolidate_models(shared_settings, models)

xtrain(model_grid, train_all = FALSE, resample_seed = 1, ...)

edit_model(model_grid, model_name, ...)

remove_model(model_grid, model_name)
}
\arguments{
\item{model_grid}{model_grid.}

\item{...}{Optional arguments.}

\item{model_name}{character, your name for a given model.}

\item{custom_control}{list, any customizations to the shared 'trControl' argument.}

\item{shared_settings}{list, settings that are shared across all models.}

\item{models}{list, list with model (specifications) from a model grid.}

\item{train_all}{logical, if TRUE train all models. If set to FALSE train only models,
for which no fit already exists.}

\item{resample_seed}{integer, this seed is used to create identical resamples across models in
order to obtain a fair comparison of the models.}
}
\description{
Train and compare multiple models with a caret model grid.
}
\details{
\describe{
  \item{add_model}{Add model to model grid.}
  \item{consolidate_models}{Consolidate models from shared settings and individual model
  specifications.}
  \item{edit_model}{Edit model within a model grid.}
  \item{model_grid}{Create empty model grid.}
  \item{remove_model}{Remove model from model grid.}
  \item{set_shared_settings}{Set shared settings for all models within a model grid.}
  \item{xtrain}{Consolidate and train model configurations from a model grid.}
}
}
