% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_model.R
\name{remove_model}
\alias{remove_model}
\title{Remove model from model grid}
\usage{
remove_model(model_grid, model_name)
}
\arguments{
\item{model_grid}{\code{model_grid}}

\item{model_name}{\code{character} with the unique name (set by the user) of
the model, you want to remove from a model grid.}
}
\value{
\code{model_grid}
}
\description{
Removes an individual model specification from a model grid. If the model has
been trained, the fitted model will also be swiped.
}
