% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model_grid.R
\name{train.model_grid}
\alias{train.model_grid}
\title{Train models within a model grid}
\usage{
\method{train}{model_grid}(mg, train_all = FALSE, resample_seed = 0)
}
\arguments{
\item{mg}{\code{model_grid}}

\item{train_all}{\code{logical} if set to TRUE, all models will be trained.
If set to FALSE only models, for which no fit already exists, will be
trained.}

\item{resample_seed}{\code{integer} is used to create identical resamples
across models in order to obtain a fair (and reproducible) comparison of
the models.}
}
\description{
Consolidates all model (and training) configurations from a model grid and
trains them with the train function from the caret package.
}
\examples{
\dontrun{
# Load dataset.
library(caret)
data(GermanCredit)

# Create model grid with a RF and XGB model.
models <-
  model_grid() \%>\%
  share_settings(
    y = GermanCredit[["Class"]],
    x = GermanCredit \%>\% select(-Class),
    metric = "ROC",
    trControl = trainControl(
      method = "cv",
      number = 5,
      summaryFunction = twoClassSummary,
      classProbs = TRUE
    )
  ) \%>\%
  add_model(
    model_name = "Funky Forest",
    method = "rf",
    tuneLength = 5
    ) \%>\%
  add_model(
    model_name = "Big Boost",
    method = "xgbTree",
    nthread = 8
    )

# Train all model configurations in model grid.
train(models)
}
}
