% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_model.R
\name{consolidate_model}
\alias{consolidate_model}
\title{Consolidate model in model grid}
\usage{
consolidate_model(shared_settings, model)
}
\arguments{
\item{shared_settings}{\code{list} settings that are shared across all models.}

\item{model}{\code{list} with the individual specifications of a model in a
model grid.}
}
\value{
\code{list} with a complete model specification, that can be trained
using the 'caret' package.
}
\description{
Consolidate model settings from shared settings and model specific settings.
In case of any overlap between the two, the model specific settings will
apply.
}
