% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_model.R
\name{edit_model}
\alias{edit_model}
\title{Edit model within a model grid}
\usage{
edit_model(model_grid, model_name, ...)
}
\arguments{
\item{model_grid}{\code{model_grid}}

\item{model_name}{\code{character} with the unique name (set by the user) of
the model, you want to modify.}

\item{...}{All the settings you want to modify for an existing model
specification.}
}
\value{
\code{model_grid}
}
\description{
Modify an existing model in a model grid.
}
