% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_grid.R
\name{model_grid}
\alias{model_grid}
\title{Pre-allocate an empty model grid}
\usage{
model_grid()
}
\value{
\code{model_grid}
}
\description{
Pre-allocate an empty model grid
}
\examples{
# Pre-allocate an empty model grid.
model_grid()
}
