% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_grid.R, R/train_model_grid.R
\name{model_grid}
\alias{model_grid}
\alias{train.model_grid}
\title{Pre-allocate an empty model grid}
\usage{
model_grid()

\method{train}{model_grid}(mg, train_all = FALSE, resample_seed = 0)
}
\arguments{
\item{mg}{\code{model_grid}}

\item{train_all}{\code{logical} if set to TRUE, all models will be trained.
If set to FALSE only models, for which no fit already exists, will be
trained.}

\item{resample_seed}{\code{integer} is used to create identical resamples
across models in order to obtain a fair (and reproducible) comparison of
the models. If set to NULL, seed will not be set (NOT advised).}
}
\value{
\code{model_grid}
}
\description{
This is a constructor function that pre-allocates an empty model grid.

The S3 method of the train function for the 'model_grid' class consolidates
all model (and training) configurations from a model grid and
trains them with the train function from the caret package.
}
\details{
The model grid makes it easy to create, manage and train multiple caret
models.
}
\examples{
# Pre-allocate an empty model grid.
model_grid()

\dontrun{
# Load dataset.
library(caret)
data(GermanCredit)

# Create model grid with a RF and XGB model.
models <-
  model_grid() \%>\%
  share_settings(
    y = GermanCredit[["Class"]],
    x = GermanCredit \%>\% dplyr::select(-Class),
    metric = "ROC",
    trControl = caret::trainControl(
      method = "cv",
      number = 5,
      summaryFunction = caret::twoClassSummary,
      classProbs = TRUE
    )
  ) \%>\%
  add_model(
    model_name = "Funky Forest",
    method = "rf",
    tuneLength = 5
    ) \%>\%
  add_model(
    model_name = "Big Boost",
    method = "xgbTree",
    nthread = 8
    )

# Train all model configurations in model grid.
train(models)
}
}
\seealso{
\code{\link{add_model}} for how to add a model to a model grid,
 \code{\link{edit_model}} for how to edit an existing model within a model grid,
 \code{\link{share_settings}} for how to define the shared settings of models
 within a model grid and
 \code{\link{remove_model}} for how to remove a model from a model grid.
}
