% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_settings.R
\name{share_settings}
\alias{share_settings}
\title{Set shared settings for a model grid}
\usage{
share_settings(model_grid, ...)
}
\arguments{
\item{model_grid}{\code{model_grid}}

\item{...}{Optional arguments.}
}
\value{
\code{model_grid} equipped with shared settings.
}
\description{
Set shared settings for all models within a model grid. These settings will
apply for a given model, unless the same settings have already been specified
in the model specific configuration. In that case the model specific settings
will apply. The 'trControl' parameter must be included in the shared
settings.
}
