% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model_grid.R
\name{train.model_grid}
\alias{train.model_grid}
\title{Train models within a model grid}
\usage{
\method{train}{model_grid}(mg, train_all = FALSE, resample_seed = 0)
}
\arguments{
\item{mg}{\code{model_grid}}

\item{train_all}{\code{logical} if TRUE train all models. If set to FALSE train only models,
for which no fit already exists.}

\item{resample_seed}{\code{integer} is used to create identical resamples across models in
order to obtain a fair (and reproducible) comparison of the models.}
}
\description{
Consolidates all model configurations from a model grid and trains them with
caret::train.
}
